HW02 Start adding allocate func
