--- HW02-annotation-conf/src/main/java/app/Attempt.java
@@ -1,7 +1,0 @@
-package app;
-
-public interface Attempt {
-    void start();
-
-    int showResult();
-}
