--- HW02-annotation-conf/src/main/java/app/MarkCalc.java
@@ -1,6 +1,0 @@
-package app;
-
-public interface MarkCalc {
-    public void addMark(int rightOptionsQuantity, boolean... answerResults);
-    public int getResult();
-}
