--- HW02-annotation-conf/src/main/java/app/OptionsShuffler.java
@@ -1,13 +1,0 @@
-package app;
-
-import java.util.Map;
-import java.util.Set;
-
-public interface OptionsShuffler {
-
-    RandomOptionsShufflerImpl shuffleAnswers(Set<String> answers);
-
-    Map<Integer, String> getSequence();
-
-    String getAnswer(int key);
-}
