--- HW02-annotation-conf/src/main/java/console/Console.java
@@ -1,6 +1,0 @@
-package console;
-
-public interface Console {
-    String read();
-    void print(String information);
-}
