--- HW02-annotation-conf/src/main/java/dao/CSVHeaders.java
@@ -1,8 +1,0 @@
-package dao;
-
-public enum CSVHeaders {
-    body,
-    type,
-    correct,
-    wrong
-}
