--- HW02-annotation-conf/src/main/java/dao/QuestionsDao.java
@@ -1,9 +1,0 @@
-package dao;
-
-import data.Question;
-
-import java.util.List;
-
-public interface QuestionsDao {
-    List<Question> loadQuestions();
-}
