--- HW02-annotation-conf/src/main/java/data/Question.java
@@ -1,12 +1,0 @@
-package data;
-
-import java.util.Set;
-
-public interface Question {
-
-    String getQuestionText();
-    Set<String> getOptions();
-    boolean checkAnswer(String answer);
-    int getQuantityOfRightOptions();
-    QuestionType getType();
-}
