--- HW02-annotation-conf/src/main/java/exceptions/AnswerException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public abstract class AnswerException extends Exception {
-    public AnswerException(String message) {
-        super(message);
-    }
-}
