--- HW02-annotation-conf/src/main/java/exceptions/EmptyAnswerException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class EmptyAnswerException extends AnswerException {
-    public EmptyAnswerException(String message) {
-        super(message);
-    }
-}
