--- HW02-annotation-conf/src/main/java/exceptions/WrongFormatAnswerException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class WrongFormatAnswerException extends AnswerException {
-    public WrongFormatAnswerException(String message) {
-        super(message);
-    }
-}
