--- HW02-annotation-conf/src/main/java/exceptions/WrongNumberAnswerException.java
@@ -1,8 +1,0 @@
-package exceptions;
-
-public class WrongNumberAnswerException extends AnswerException {
-
-    public WrongNumberAnswerException(String message) {
-        super(message);
-    }
-}
