+++ HW02-annotation-conf/src/main/java/ru/otus/hw02/app/Attempt.java
@@ -1,0 +1,7 @@
+package ru.otus.hw02.app;
+
+public interface Attempt {
+    void start();
+
+    int showResult();
+}
