+++ HW02-annotation-conf/src/main/java/ru/otus/hw02/app/MarkCalc.java
@@ -1,0 +1,6 @@
+package ru.otus.hw02.app;
+
+public interface MarkCalc {
+    public void addMark(int rightOptionsQuantity, boolean... answerResults);
+    public int getResult();
+}
