+++ HW02-annotation-conf/src/main/java/ru/otus/hw02/app/OptionsShuffler.java
@@ -1,0 +1,13 @@
+package ru.otus.hw02.app;
+
+import java.util.Map;
+import java.util.Set;
+
+public interface OptionsShuffler {
+
+    RandomOptionsShufflerImpl shuffleAnswers(Set<String> answers);
+
+    Map<Integer, String> getSequence();
+
+    String getAnswer(int key);
+}
