+++ HW02-annotation-conf/src/main/java/ru/otus/hw02/console/Console.java
@@ -1,0 +1,6 @@
+package ru.otus.hw02.console;
+
+public interface Console {
+    String read();
+    void print(String information);
+}
