+++ HW02-annotation-conf/src/main/java/ru/otus/hw02/console/ConsoleImpl.java
@@ -1,0 +1,28 @@
+package ru.otus.hw02.console;
+
+import org.springframework.stereotype.Service;
+
+import java.util.Objects;
+import java.util.Scanner;
+
+
+@Service
+public class ConsoleImpl implements Console {
+    private final Scanner scanner;
+
+    public ConsoleImpl() {
+        scanner = new Scanner(System.in);
+    }
+
+    @Override
+    public String read() {
+        return scanner.nextLine();
+    }
+
+    @Override
+    public void print(String information) {
+        if (Objects.nonNull(information)) {
+            System.out.print(information);
+        }
+    }
+}
