+++ HW02-annotation-conf/src/main/java/ru/otus/hw02/dao/CSVHeaders.java
@@ -1,0 +1,8 @@
+package ru.otus.hw02.dao;
+
+public enum CSVHeaders {
+    body,
+    type,
+    correct,
+    wrong
+}
