+++ HW02-annotation-conf/src/main/java/ru/otus/hw02/dao/QuestionsDao.java
@@ -1,0 +1,10 @@
+package ru.otus.hw02.dao;
+
+import ru.otus.hw02.data.Question;
+
+import java.io.IOException;
+import java.util.List;
+
+public interface QuestionsDao {
+    List<Question> loadQuestions() throws IOException;
+}
