+++ HW02-annotation-conf/src/main/java/ru/otus/hw02/data/Question.java
@@ -1,0 +1,12 @@
+package ru.otus.hw02.data;
+
+import java.util.Set;
+
+public interface Question {
+
+    String getQuestionText();
+    Set<String> getOptions();
+    boolean checkAnswer(String answer);
+    int getQuantityOfRightOptions();
+    QuestionType getType();
+}
