+++ HW02-annotation-conf/src/main/java/ru/otus/hw02/data/QuestionType.java
@@ -1,0 +1,17 @@
+package ru.otus.hw02.data;
+
+public enum QuestionType {
+    SIMPLE ("simple"),
+    MULTIPLE ("multiple"),
+    OPEN ("open");
+
+    private final String type;
+
+    QuestionType(String type) {
+        this.type = type;
+    }
+
+    public String getType() {
+        return type;
+    }
+}
