+++ HW02-annotation-conf/src/main/java/ru/otus/hw02/exceptions/AnswerException.java
@@ -1,0 +1,7 @@
+package ru.otus.hw02.exceptions;
+
+public abstract class AnswerException extends Exception {
+    public AnswerException(String message) {
+        super(message);
+    }
+}
