+++ HW02-annotation-conf/src/main/java/ru/otus/hw02/exceptions/EmptyAnswerException.java
@@ -1,0 +1,7 @@
+package ru.otus.hw02.exceptions;
+
+public class EmptyAnswerException extends AnswerException {
+    public EmptyAnswerException(String message) {
+        super(message);
+    }
+}
