+++ HW02-annotation-conf/src/main/java/ru/otus/hw02/exceptions/WrongFormatAnswerException.java
@@ -1,0 +1,7 @@
+package ru.otus.hw02.exceptions;
+
+public class WrongFormatAnswerException extends AnswerException {
+    public WrongFormatAnswerException(String message) {
+        super(message);
+    }
+}
