+++ HW02-annotation-conf/src/main/java/ru/otus/hw02/exceptions/WrongNumberAnswerException.java
@@ -1,0 +1,8 @@
+package ru.otus.hw02.exceptions;
+
+public class WrongNumberAnswerException extends AnswerException {
+
+    public WrongNumberAnswerException(String message) {
+        super(message);
+    }
+}
