+++ HW02-annotation-conf/src/main/resources/application.properties
@@ -1,1 +1,2 @@
 db.url = src/main/resources/questions.csv
+offset = 50
