+++ HW02-annotation-conf/src/main/resources/il8n/bundle_en_EN.properties
@@ -1,0 +1,2 @@
+print.name = Print your name please: 
+print.surname = Print your surname:
