+++ HW02-annotation-conf/src/main/resources/il8n/bundle_ru_RU.properties
@@ -1,0 +1,2 @@
+print.name=Напишите своё имя: 
+print.surname=Напишите свою фамилию:
