--- HW02-annotation-conf/src/test/java/console/ConsoleTest.java
@@ -1,28 +1,0 @@
-package console;
-
-import org.junit.jupiter.api.DisplayName;
-import org.junit.jupiter.api.Test;
-
-import java.util.Scanner;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-@DisplayName("Testing Console.class")
-public class ConsoleTest {
-
-    @DisplayName("Checking that class Console is correctly created")
-    @Test
-    void shouldCorrectCreated() {
-        Console console = new ConsoleImpl();
-        assertNotNull(console);
-    }
-
-    @DisplayName("Testing correct printing")
-    @Test
-    void shouldCorrectPrint() {
-//        Scanner scanner = new Scanner(System.in);
-//        scanner.nextLine();
-//        System.
-
-    }
-}
